Break up Core Components
