+++ lib/job_index_web/components/core/form/label.ex
@@ -1,0 +1,19 @@
+defmodule JinWeb.Components.Core.Form.Label do
+  @moduledoc false
+
+  use Phoenix.Component
+
+  @doc """
+  Renders a label.
+  """
+  attr :for, :string, default: nil
+  slot :inner_block, required: true
+
+  def label(assigns) do
+    ~H"""
+    <label for={@for} class="block text-sm font-semibold leading-6 text-zinc-800">
+      <%= render_slot(@inner_block) %>
+    </label>
+    """
+  end
+end
